requirements.txt
Flask
Flask-CORS
python-dotenv
requests
openai
beautifulsoup4
lxml
selenium
webdriver-manager
pandas
numpy
matplotlib
seaborn
duckdb